requirements.txt
requests==2.31.0
python-telegram-bot==20.8
python-dotenv==1.0.0
schedule==1.2.0
urllib3==2.0.7